# Conflicts:
#	package-lock.json
#	package.json

+++ utils/index.js
@@ -1,5 +1,5 @@
 'use strict';
-const generators = require('yeoman-generator');
+const Generator = require('yeoman-generator');
 const {merge, template} = require('lodash');
 const path = require('path');
 const glob = require('glob').sync;
@@ -74,11 +74,21 @@ function writeTemplates(config, withSamples) {
   };
 
   const copy = (prefix) => {
-    this.fs.copy(this.templatePath(prefix + 'plain/**/*'), this.destinationPath(), includeDot);
+    this.fs.exists(this.templatePath(prefix + 'plain'), (exists) => {
+      if(exists === false) {
+        return;
+      }
+      this.fs.copy(this.templatePath(prefix + 'plain/**/*'), this.destinationPath(), includeDot);
+    });
     copyTpl(this.templatePath(prefix + 'processed'), '');
 
     if (config.name) {
-      this.fs.copy(this.templatePath(prefix + 'pluginname_plain/**/*'), this.destinationPath(config.name.toLowerCase() + '/'), includeDot);
+      this.fs.exists(this.templatePath(prefix + 'pluginname_plain'), (exists) => {
+        if(exists === false) {
+          return;
+        }
+        this.fs.copy(this.templatePath(prefix + 'pluginname_plain/**/*'), this.destinationPath(config.name.toLowerCase() + '/'), includeDot);
+      });
       copyTpl(this.templatePath(prefix + 'pluginname_processed'), config.name.toLowerCase() + '/');
     }
   };
@@ -96,7 +106,7 @@ function useDevVersion() {
   return (pkg.version || '').includes('-');
 }
 
-class BaseInitPluginGenerator extends generators.Base {
+class BaseInitPluginGenerator extends Generator {
 
   constructor(args, options, basetype) {
     super(args, options);
@@ -139,7 +149,7 @@ class BaseInitPluginGenerator extends generators.Base {
       this._patchPackageJSON(config);
     }
     if (this.fs.exists(this.templatePath('_gitignore'))) {
-      this.copy(this.templatePath('_gitignore'), this.destinationPath('.gitignore'));
+      this.fs.copy(this.templatePath('_gitignore'), this.destinationPath('.gitignore'));
     }
 
     this._writeTemplates(config, !this.options.noSamples);
